chore: break out ConfirmSummary into a separate component
